+++ APExtensions/Classes/Storyboard/UIScrollView+Storyboard.swift
@@ -8,25 +8,131 @@
 
 import UIKit
 
+//-----------------------------------------------------------------------------
+// MARK: - Helper Extension
+//-----------------------------------------------------------------------------
 
+fileprivate extension UIView {
+    fileprivate var _viewController: UIViewController? {
+        var nextResponder: UIResponder? = self
+        while nextResponder != nil {
+            nextResponder = nextResponder?.next
+            
+            if let viewController = nextResponder as? UIViewController {
+                return viewController
+            }
+        }
+        
+        return nil
+    }
+}
+
+//-----------------------------------------------------------------------------
+// MARK: - Adjustment Behavior
+//-----------------------------------------------------------------------------
+
+@available(iOS 11.0, *)
 public extension UIScrollView {
-    /// Sets top 64 for inset
-    @IBInspectable public var avoidTopBars: Bool {
+    @IBInspectable public var disableAutomaticContentAdjustment: Bool {
         get {
-            return contentInset.top == 64
+            return contentInsetAdjustmentBehavior == .never
         }
         set {
-            contentInset.top = 64
+            contentInsetAdjustmentBehavior = newValue ? .never : .automatic
+        }
+    }
+}
+
+//-----------------------------------------------------------------------------
+// MARK: - Bars Avoid
+//-----------------------------------------------------------------------------
+
+public extension UIScrollView {
+    @available(iOS, deprecated: 11.0, obsoleted: 12.0, message: "avoidTopBars was deprecated, please use avoidNavigationBar")
+    @IBInspectable public var avoidTopBars: Bool {
+        get {
+            NSLog("avoidTopBars was deprecated, please use avoidNavigationBar")
+            
+            return avoidNavigationBar
+        }
+        set {
+            NSLog("avoidTopBars was deprecated, please use avoidNavigationBar")
+            
+            avoidNavigationBar = newValue
         }
     }
     
-    /// Sets 49 for bottom inset
-    @IBInspectable public var avoidTabBar: Bool {
+    /// Sets 64 or 0 for top content inset and disables automatic mechanisms to prevent conflict.
+    /// Returns true if scroll view avoids top bars. Takes into account `contentInsetAdjustmentBehavior`.
+    @IBInspectable public var avoidNavigationBar: Bool {
         get {
-            return contentInset.top == 49
+            if #available(iOS 11.0, *) {
+                switch contentInsetAdjustmentBehavior {
+                case .always: return adjustedContentInset.top == 64
+                case .never: return contentInset.top == 64
+                    
+                case .scrollableAxes:
+                    if isScrollEnabled || alwaysBounceVertical {
+                        return adjustedContentInset.top == 64
+                    } else {
+                        return contentInset.top == 64
+                    }
+                
+                case .automatic:
+                    if let _ = _viewController?.navigationController {
+                        return adjustedContentInset.top == 64
+                    } else {
+                        return contentInset.top == 64
+                    }
+                }
+            } else {
+                return contentInset.top == 64
+            }
         }
         set {
-            contentInset.bottom = 49
+            if #available(iOS 11.0, *) {
+                disableAutomaticContentAdjustment = true
+            }
+            
+            _viewController?.automaticallyAdjustsScrollViewInsets = false
+            contentInset.top = newValue ? 64 : 0
+        }
+    }
+    
+    /// Sets 49 or 0 for bottom inset and disables automatic mechanisms to prevent conflict.
+    /// Returns true if scroll view avoids bottom bars. Takes into account `contentInsetAdjustmentBehavior`.
+    @IBInspectable public var avoidTabBar: Bool {
+        get {
+            if #available(iOS 11.0, *) {
+                switch contentInsetAdjustmentBehavior {
+                case .always: return adjustedContentInset.bottom == 49
+                case .never: return contentInset.bottom == 49
+                    
+                case .scrollableAxes:
+                    if isScrollEnabled || alwaysBounceVertical {
+                        return adjustedContentInset.bottom == 49
+                    } else {
+                        return contentInset.bottom == 49
+                    }
+                    
+                case .automatic:
+                    if let _ = _viewController?.navigationController {
+                        return adjustedContentInset.bottom == 49
+                    } else {
+                        return contentInset.bottom == 49
+                    }
+                }
+            } else {
+                return contentInset.bottom == 49
+            }
+        }
+        set {
+            if #available(iOS 11.0, *) {
+                disableAutomaticContentAdjustment = true
+            }
+            
+            _viewController?.automaticallyAdjustsScrollViewInsets = false
+            contentInset.bottom = newValue ? 49 : 0
         }
     }
 }