+++ Example/APExtensions/ViewController.swift
@@ -11,6 +11,8 @@ import APExtensions
 
 
 class ViewController: UIViewController {
+    
+    @IBOutlet private weak var scrollView: UIScrollView!
 
     //-----------------------------------------------------------------------------
     // MARK: - UIViewController Overrides
@@ -19,14 +21,22 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        // Get all classes that conform to SetupOnce protocol
+        let setupOnces: [SetupOnce.Type] = g_getClassesConformToProtocol(SetupOnce.self)
+        print("Classes conform to SetupOnce: \(setupOnces)")
+        
         print(Date().previousWorkDay)
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        
+        
     }
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         
-        // Get all classes that conform to SetupOnce protocol
-        let setupOnces: [SetupOnce.Type] = g_getClassesConformToProtocol(SetupOnce.self)
-        print("Classes conform to SetupOnce: \(setupOnces)")
+        
     }
 }